Add native builds, static feature and BLOSC2_INSTALL_PREFIX (#8)
Add BLOSC2_INSTALL_PREFIX, when used w/ --use-system-blosc2, it will indicate 
where to look for the blosc2 install. Without it, fallback to trying to locate with
pkg-config. When not used with --use-system-blosc2, it will dictate where to install
blosc2 during build, defaulting to the OUT_DIR.

Add --static to indicate to try and statically link to blosc2.

Add native builds along w/ conda builds to ensure both work and also just
figure out what is needed when building in and out of a conda environment.